% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{geoMeanOfOddsCalc}
\alias{geoMeanOfOddsCalc}
\title{Geometric Mean of Odds}
\usage{
geoMeanOfOddsCalc(x, q = 0.05, odds = FALSE)
}
\arguments{
\item{x}{A vector of forecasts (probabilities! unless odds = TRUE)}

\item{q}{The quantile to use for replacing 0s (between 0 and 1)}

\item{odds}{Whether x is already in odds form (TRUE) or probabilities}
}
\value{
(numeric) The geometric mean of the odds
}
\description{
Convert probabilities to odds, and calculate the geometric mean of the
odds. We handle 0s by replacing them with the qth quantile of the non-zero
forecasts, before converting.
}
\note{
agg(a) + agg(not a) does not sum to 1 for this aggregation method.
}
