% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{hd_trim}
\alias{hd_trim}
\title{Highest-Density Trimmed Mean}
\usage{
hd_trim(x, p = 0.1)
}
\arguments{
\item{x}{Vector of forecasts in 0 to 100 range (\%)}

\item{p}{The proportion of forecasts to trim (between 0 and 1)}
}
\value{
(numeric) The highest-density trimmed mean of the vector
}
\description{
From Powell et al. (2022) \doi{10.1037/dec0000191}. You find
the shortest interval containing (1-p) * 100\% of the data and take the mean
of the forecasts within that interval.
}
\note{
As p gets bigger this acts like a mode in a similar way to
the symmetrically-trimmed mean acting like a median.
}
