% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{neymanAggCalc}
\alias{neymanAggCalc}
\title{Neyman Aggregation (Extremized)}
\usage{
neymanAggCalc(x)
}
\arguments{
\item{x}{Vector of forecasts in 0 to 100 range (\%)}
}
\value{
(numeric) The extremized mean of the vector
}
\description{
Takes the arithmetic mean of the log odds of the forecasts,
then extremizes the mean by a factor d, where d is

(n*(sqrt((3\emph{n^2) - (3}n) + 1) - 2))/(n^2 - n - 1)

where n is the number of forecasts.
}
\references{
Neyman, E. and Roughgarden, T. (2021). Are you
smarter than a random expert? The robust aggregation of substitutable
signals: \doi{10.1145/3490486.3538243}. Also Jaime Sevilla's EAF post
``Principled extremizing of aggregated forecasts."
}
