% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{trim}
\alias{trim}
\title{Trimmed mean}
\usage{
trim(x, p = 0.1)
}
\arguments{
\item{x}{Vector of forecasts in 0 to 100 range (\%)}

\item{p}{The proportion of forecasts to trim from each end (between 0 and
1)}
}
\value{
(numeric) The trimmed mean of the vector
}
\description{
Trim the top and bottom (p*100)\% of forecasts
}
